sussa-api/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.1.7
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
dnspython==2.3.0
eventlet==0.33.3
fcm-django==1.0.12
firebase-admin==5.4.0
flower==1.2.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.5
Pillow==9.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
pytz==2022.7.1
redis==4.5.4
requests==2.29.0
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.3.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
